+++ gulpTasks/testing.js
@@ -1,0 +1,14 @@
+const gulp = require('gulp');
+const mocha = require('gulp-spawn-mocha');
+const options = require('../gulpfile.js').options;
+
+
+gulp.task('test', () => {
+    return gulp.src([
+        `./tests/${options.type}/*.test.js`,
+    ]).pipe(mocha({
+        timeout: 60000,
+        ui: 'exports',
+        reporter: 'spec',
+    }));
+});
