+++ package.json
@@ -55,10 +55,12 @@
     "gulp": "^3.9.0",
     "gulp-spawn-mocha": "^3.1.0",
     "istanbul": "^0.4.5",
+    "json-structure-diff": "^0.0.2",
     "merge-stream": "^1.0.0",
     "minimist": "^1.2.0",
     "mocha": "^3.2.0",
     "optimist": "^0.6.1",
+    "require-dir": "^0.3.1",
     "run-sequence": "^1.2.1",
     "semver-compare": "^1.0.0",
     "shelljs": "^0.7.6",
